+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarAutocompleteFeatureTest.kt
@@ -115,6 +115,14 @@ class ToolbarAutocompleteFeatureTest {
             fail()
         }
 
+        override fun hideMenuButton() {
+            fail()
+        }
+
+        override fun showMenuButton() {
+            fail()
+        }
+
         override fun invalidateActions() {
             fail()
         }