+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarInteractorTest.kt
@@ -98,6 +98,14 @@ class ToolbarInteractorTest {
             fail()
         }
 
+        override fun hideMenuButton() {
+            fail()
+        }
+
+        override fun showMenuButton() {
+            fail()
+        }
+
         override fun invalidateActions() {
             fail()
         }